all cruds implemented

+++ service/src/main/java/com/epam/lab/service/NewsService.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface NewsService extends CrudService<News> {
     void addAuthor(News news, Author author);
     void addTags(List<Tag> tags);
-    int countNews();
+    long countNews();
     List<News> readNews(SearchCriteria criteria);
 }
